feat(translation): add Esperanto translation (#471)
* Added eo

* 1/3 done translating to Esperanto

* Added esperanto

The esperanto flag has a star, but no emoji yet (we are working on this though!)

* Finished translations of esperanto

There's no direct for the concept of "forked", but forkis would logistically make the most since in the language. If there ever becomes a consensus on this, change it then. Otherwise I'm coining it now.

* Added esperanto

* Fixed typo

* Fixed flag